Added file example code
