+++ docs/current_docs/manuals/developer/snippets/functions/arguments-context-dir-file/go/.gitattributes
