+++ docs/current_docs/manuals/developer/snippets/functions/arguments-context-dir-file/python/main.py
@@ -11,3 +11,10 @@ class MyModule:
         source: Annotated[dagger.Directory, DefaultPath("/")],
     ) -> list[str]:
         return await source.entries()
+
+    @function
+    async def readFile(
+        self,
+        source: Annotated[dagger.File, DefaultPath("./README.md")],
+    ) -> list[str]:
+        return await source.contents()
