+++ docs/current_docs/manuals/developer/snippets/functions/arguments-context-dir-file/typescript/index.ts
@@ -1,0 +1,18 @@
+import { Directory, File, argument, object, func } from "@dagger.io/dagger"
+
+@object()
+class MyModule {
+  @func()
+  async readDir(
+    @argument({ defaultPath: "/" }) source: Directory,
+  ): Promise<string[]> {
+    return await source.entries()
+  }
+
+  @func()
+  async readFile(
+    @argument({ defaultPath: "/README.md" }) source: File,
+  ): Promise<string> {
+    return await source.contents()
+  }
+}
