--- docs/current_docs/manuals/developer/snippets/functions/arguments-contextual-dir/typescript/index.ts
@@ -1,11 +1,0 @@
-import { dag, Directory, File, argument, object, func } from "@dagger.io/dagger"
-
-@object()
-class MyModule {
-  @func()
-  async readDir(
-    @argument({ defaultPath: "/" }) source: Directory,
-  ): Promise<string[]> {
-    return await source.entries()
-  }
-}
